+++ src/main/java/suite/math/Tanh.java
@@ -5,23 +5,23 @@ public class Tanh {
 	public static float[] tanhOn(float[] fs) {
 		int length = fs.length;
 		for (int i = 0; i < length; i++)
-			fs[i] = tanh(fs[i]);
+			fs[i] = (float) tanh(fs[i]);
 		return fs;
 	}
 
 	public static float[] tanhGradientOn(float[] fs) {
 		int length = fs.length;
 		for (int i = 0; i < length; i++)
-			fs[i] = tanhGradient(fs[i]);
+			fs[i] = (float) tanhGradient(fs[i]);
 		return fs;
 	}
 
-	public static float tanh(float value) {
-		return (float) Math.tanh(value);
+	public static double tanh(double value) {
+		return Math.tanh(value);
 	}
 
-	public static float tanhGradient(float value) {
-		return 1f - value * value;
+	public static double tanhGradient(double value) {
+		return 1d - value * value;
 	}
 
 }
